utils, group request expiration, GroupRequestQueue uses Round
A new Utils class for common utility functions.
Currently contains functions for converting between datetime formats.

Group requests now have an expiration time (enforced in the queue and by the driver),
after which they are deleted.
The expiration time is set for the game in the config file as group_wait_time.

GroupRequeustQueue uses the Round of a request to look for matches.
(Note that the field is still called "step" in the database.)