exit trail implementation
When a session times out, or a partner's session times out, or something else happens that means we can't continue the experiment,
the sessions (if they are still polling) get placed on an "exit trail":
a sequence of steps ("ExistStep") that behave (mostly) like regular steps.
That way, for example, the experimenter can collect demographic data from the users, so not everything is lost -- or at least display a nice exit message before returning.